Refactor Send/Receive MOB page
