+++ app/pages/Contacts/ContactsPage.presenter/ContactsPage.presenter.tsx
@@ -117,3 +117,4 @@ const ContactsPage: FC = () => {
 };
 
 export default ContactsPage;
+export { ContactsPage };
