+++ app/pages/SendReceive/ReceiveMob.view/ReceiveMob.d.ts
@@ -1,0 +1,7 @@
+import type Account from '../../../types/Account';
+import type Contact from '../../../types/Contact';
+
+export interface ReceiveMobProps {
+  contacts: Contact[];
+  selectedAccount: Account;
+}
