+++ app/pages/SendReceive/SendMob.view/index.ts
@@ -1,0 +1,4 @@
+import { SendMob } from './SendMob.view';
+
+export default SendMob;
+export { SendMob };
