+++ app/pages/SendReceive/index.ts
@@ -1,0 +1,4 @@
+import { SendReceivePage } from './SendReceivePage.presenter';
+
+export default SendReceivePage;
+export { SendReceivePage };
