+++ app/routes.tsx
@@ -7,7 +7,7 @@ import UnlockWalletGuard from './components/UnlockWalletGuard';
 import WalletGuard from './components/WalletGuard';
 import routePaths from './constants/routePaths';
 import DashboardLayout from './layouts/DashboardLayout';
-import { ContactsPage } from './pages';
+import { ContactsPage, SendReceivePage } from './pages';
 import { CreateAccountView, ImportAccountView, UnlockWalletView } from './views/auth';
 import NotFoundView from './views/errors/NotFoundView';
 import {
@@ -21,7 +21,6 @@ import {
   SettingsView,
   PrivacyPolicyView,
   TermsOfUseView,
-  TransactionView,
 } from './views/wallet';
 
 type Routes = {
@@ -100,7 +99,7 @@ const routes: Routes = [
         path: routePaths.APP_DASHBOARD,
       },
       {
-        Component: TransactionView,
+        Component: SendReceivePage,
         exact: true,
         path: routePaths.APP_TRANSACTION,
       },
