--- app/views/wallet/TransactionView/SendMobPanel/index.ts
@@ -1,3 +1,0 @@
-import SendMobPanel from './SendMobPanel';
-
-export default SendMobPanel;
